pyproject.toml
[project]
name = "churn-mlops-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.39.2",
    "botocore>=1.39.2",
    "chardet>=5.2.0",
    "dagshub>=0.5.10",
    "dvc>=3.60.1",
    "ensure>=1.0.4",
    "fastapi>=0.115.14",
    "httpx>=0.28.1",
    "imbalanced-learn>=0.13.0",
    "joblib>=1.5.1",
    "langchain-core>=0.3.67",
    "matplotlib>=3.10.3",
    "mlflow==2.2.2",
    "numpy>=1.26.4",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pydantic>=2.11.7",
    "python-box>=7.3.2",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "streamlit>=1.46.1",
    "typing-extensions>=4.14.0",
    "uvicorn>=0.35.0",
    "xlrd>=2.0.2",
]
